+++ xclim/core/utils.py
@@ -383,7 +383,7 @@ def raise_warn_or_log(
     elif mode == "warn":
         warnings.warn(msg, stacklevel=stacklevel + 1)
     else:  # mode == "raise"
-        raise err
+        raise err from ValueError(msg)
 
 
 class InputKind(IntEnum):