+++ HISTORY.rst
@@ -9,6 +9,7 @@ Contributors to this version: Pascal Bourgault (:user:`aulemahal`), Trevor James
 New features and enhancements
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 * Improve cell_methods checking to search the wanted method within the whole string. (:pull:`866`, :issue:`863`).
+* New ``align_on='random`` option for ``xclim.core.calendar.convert_calendar``, for conversions involving '360_day' calendars. (:pull:`875`, :issue:`841`).
 
 Breaking changes
 ~~~~~~~~~~~~~~~~